V2 dev (#33)
* Added timeline styles

* Added timeline and placeholder calls

* Adjusted timeline styles with panels

Adjust timeline when panels option object is set to true.

* Cleaned up code and added comments

Added comments and cleaned up code for readability.

* Removed console.log calls

Fixed tabbable icon showing if time-line is enabled on row.

* Darken borders to timeline blocks.

The borders were too light for the blocks in the timeline. Darken the borders a bit.

* Added timeline and removed themes

Removed themes and added layout option to use box option.

* Remove themes

Removed themes.

* Fixed tabbable

Fixed where the tabbable toggle switches were not appearing when selecting additional columns.

* Add lazyload feature

Added lazyload feature.

* Added lazyload styles

Added styles for new lazyload feature.

* Fixed Lazyload when Searchable is also enabled.

Set "Load more" button to hide when TypeAhead, Searchable is also enabled on the People panels.

* Fixed Load More button in tab-panes

Added bottom margining to the "Load More" button when LazyLoad is enabled and the people panels are set in tab containers.

* Update package.json

* Update README.md

* Added Get Involved section

Added instructions on how to contribute to code.